Hack for Win32 performance issues.  It seems SDL_UpdateRects() really kills performance in Windows, so we use SDL_Flip() instead, but only when something has actually changed onscreen.  At some point after the next release, I may move to hardware/double-buffered displays, but the extra work it requires (dirty rect list for each buffer) is just too much for now.  Besides, I'm hesitant to spend so much time optimizing for one specific type of software mode when OpenGL is available, and works much faster anyway.
Fixed bug with pathnames in FSNodeWin32.  The code now assumes that
all paths start with X: (where X is a drive letter).  If not, then
it defaults to C:  So that means relative directories aren't allowed,
only absolute.

Removed hardware support from FrameBufferSoft, since the only port
that uses it is PSP, and that port will soon get its own
FrameBufferPSP class.  So for now, PSP support is broken.  It was
just too difficult to merge the common FrameBufferSoft for PSP
support, since it's starting to use PSP-specific code.