Removed all reference to 'user.pro', in favour of the previous (to v/2.0) functionality.  The new scheme is as follows:  -  Per-user properties are stored in stella.pro, which is located in     system-specific directories (homedir for UNIX/POSIX systems, currentdir     for Win32 systems).  -  Completely removed concept of a global properties file, since that     functionality is now included in Stella directly.  -  Retained the ability to use an alternative properties file, in which     case the per-user one isn't used at all, and any entries in it override     the built-in defaults.
Removed all reference to 'system' vs. 'user' config files, for much the
same reasons as above.  As above, config files are stored in
user-specific places depending on platform.  There is no longer a
system-wide config file.  This made sense when Stella didn't have built-in
defaults and relied on such a file, but that's no longer the case.

Sorry for any confusion this causes, but it was causing even more confusion
for users.  And now we can distribute Stella has a single binary which
doesn't depend on any external files (but will still use them if they're
present).