Moved AtariVox support from experimental to full-fledged controller.
Made SpeakJet emulation be optional, because it's still a WIP and we'll
soon be able to have AVox support without it.

Added SerialPort infrastructure, which will implement serial port access
in a general way.  Those ports which don't wish to do so can simply not
implement a SerialPortXXX class; in that case a null SerialPort is used.

Still TODO is actually implement the serial port functionality for UNIX,
and add similar classes for OSX and Win32.