Reworked 'fullres' argument to also accept the 'auto' option.  In this case, fullscreen resolutions will be automatically chosen based on the required size for the window.  The image will be centered and keep the same aspect ratio, however, so operation will still work correctly on widescreen monitors.  'Auto' will be the new default.  Otherwise, if a specific resolution is requested, Stella will try to accomodate it *only* if it fits into the resolution; otherwise the smallest resolution that fits will be used.
Removed 'zoom_ui' and 'zoom_tia'.  The UI can now only be at 1x mode.

Aded 'tia_filter' commandline argument, which specifies to the filter
to use when rendering the tia image.  For now, these accept 'zoom1x',
'zoom2x'..., up to 'zoom10x', and duplicate previous behaviour.  Eventually,
Scalexx and HQxx filters may be added.  Still TODO is add this to the UI.

First pass at making the standard build use a minimum of zoom2x for the TIA,
so the UI can be larger and use a better looking font.  There's still work
to do in this area, especially for those ports with limited hardware that
support zoom1x only.