Fixed issue with incorrect reporting of OpenGL version number for Windows. Because the version was always being returned as 0, the TV effects were always disabled, even on hardware that could support them.
Creation of OpenGL video mode is now more robust.  If it fails for any
reason, fallback to software mode is now done (and a message stating
that it happened).  This still won't protect against the operating
system lying to Stella and saying OpenGL support is really available
when it isn't, though.

Added support for ROMs smaller than 2K.  These are still treated as
2K ROMs, but have their contents mirrored appropriately.

Bumped version number.