Of course, we're still not ready for the 3.1 release, so I have to bump it again.
Rant mode : each release of Stella seems to get harder and harder.
I have to fight to force myself to get ready for release time,
and when it comes, there always seems to be something popping
up at the last minute.  This time, it's crashes on various
iMacs that I just got access to (while everything is working
perfectly on my Macbook Pro).  I sometimes think it's time
to throw in the towel, and move on to other things ...