Reworked M6502 detection of debugger GFX sections, by moving the actual detection and marking of those areas into the TIA class itself, specifically at the location of the write.  In other words, there's no longer an IF statement executed for every STx opcode; the marking is done directly within the write to GRPx or PFx, so extra code is only executed when actually storing to those locations.
Fixed several cases of opcodes marking an area as CODE when it should have
been DATA.

Added output to the CPU area of the debugger for displaying the source
address for loading data into the A/X/Y registers.  Note that these are
only modified when actual addresses are used, so immediate and zero-page
mode will show addresses as zero (meaning that no address was involved
in retrieving the data).

Tweaked console font to better diffentiate the disassembly output between
graphics for players and graphics for the playfield.

Bumped state file format because of changes to M6502.  This means old
state files will be broken.

Bumped version # for next test release.