Added logging capability and viewer.  This is useful for those platforms that don't normally use the commandline (mostly Windows, but in some cases OSX as well).  The 'showinfo' commandline argument has been renamed 'loglevel', but it has the same purpose.  A new option 'logtoconsole' has been added, which determines whether log output should also be directed to the commandline/ console (previously, it was always printed to the console).  All these items are now accessible from Options -> System Logs.
For anyone reading this (and that cares), now I can finally move on to the
OpenGL rewrite.  The plan is that the new code will use OpenGL ES, which
is a subset of OpenGL 1.5.  The main advantages are that you won't need
an advanced OpenGL card, and OpenGL ES is supported on most new 'smaller'
systems (iPhone, Android, etc), making ports much easier.